refactor: [FFBV-127] 상품 상세정보 컴포넌트 분리
