+++ src/pages/products/ProductsDetailPage.vue
@@ -3,8 +3,10 @@ import { ref, onMounted, computed } from 'vue';
 import { useRoute } from 'vue-router';
 import { getProductDetail, addScrap, delScrap } from '@/api/product-detail';
 
-import BaseCard from '@/components/common/BaseCard.vue';
-import BaseButton from '@/components/common/BaseButton.vue';
+import ProductHeader from '@/components/products/ProductHeader.vue';
+import ProductSummaryCard from '@/components/products/ProductSummaryCard.vue';
+import ProductAISummary from '@/components/products/ProductAISummary.vue';
+import ProductInfo from '@/components/products/ProductInfo.vue';
 import ProfitCalculator from '@/components/products/ProfitCalculator.vue';
 
 const productDetail = ref(null);
@@ -14,102 +16,12 @@ const renewDate = '25.06.20';
 
 const route = useRoute();
 
-const idPrefix = computed(() => {
-  if (
-    productDetail.value &&
-    productDetail.value.id &&
-    typeof productDetail.value.id === 'string'
-  ) {
-    return productDetail.value.id.charAt(0).toUpperCase();
-  }
-  return '';
-});
-
-const productTypeName = computed(() => {
-  const typeMapping = {
-    D: '예금',
-    S: '적금',
-    A: '공격형 연금',
-    C: '안정형 연금',
-    F: '펀드',
-    X: '외화',
-  };
-  return typeMapping[idPrefix.value] || '금융상품';
-});
-
-// prefix별 계산기 설정
-const calculatorProps = computed(() => {
-  if (!productDetail.value) return {};
-
-  const prefix = idPrefix.value;
-
-  switch (prefix) {
-    case 'F': // 펀드
-      return {
-        rateOptions: [
-          {
-            label: '기본 수익률',
-            value: parseFloat(productDetail.value.rate) || 0,
-          },
-          {
-            label: '기본 수익률',
-            value: parseFloat(productDetail.value.rate) || 0,
-          }, // 펀드는 primeRate가 rate와 같음
-        ],
-        period: 12, // 펀드는 기본 12개월
-        minAmount: parseFloat(productDetail.value.minimumCost) || 10000,
-      };
-    case 'X': // 외화
-      return {
-        rateOptions: [
-          {
-            label: '기본 금리',
-            value: parseFloat(productDetail.value.rate) || 0,
-          },
-          {
-            label: '기본 금리',
-            value: parseFloat(productDetail.value.rate) || 0,
-          }, // 외화도 동일
-        ],
-        period: 12, // 자동연장 기간은 숫자가 아님
-        minAmount: parseFloat(productDetail.value.minimumCost) || 10000,
-      };
-    case 'P': // 연금
-      return {
-        rateOptions: [
-          {
-            label: '기본 금리',
-            value: parseFloat(productDetail.value.rate) || 0,
-          },
-          {
-            label: '전년도 수익률',
-            value: parseFloat(productDetail.value.primeRate) || 0,
-          },
-        ],
-        period: 12, // 연금 타입은 문자열
-        minAmount: parseFloat(productDetail.value.minimumCost) || 10000,
-      };
-    default: // D, S (예금, 적금)
-      return {
-        rateOptions: [
-          {
-            label: '일반 금리',
-            value: parseFloat(productDetail.value.rate) || 0,
-          },
-          {
-            label: '최대 금리',
-            value: parseFloat(productDetail.value.primeRate) || 0,
-          },
-        ],
-        period: parseInt(productDetail.value.saveTerm) || 12,
-        minAmount: parseFloat(productDetail.value.minimumCost) || 10000,
-      };
-  }
-});
-
 const navigateToLink = () => {
-  //   window.location.href = productDetail.value.regLink;
   window.open(productDetail.value.regLink, '_blank');
+};
+
+const handleAddScrap = (productId) => {
+  addScrap(productId);
 };
 
 onMounted(async () => {
@@ -138,207 +50,28 @@ onMounted(async () => {
     <div v-else-if="productDetail">
       <!-- <div v-if="idPrefix === 'D'"> -->
       <!-- 헤더 섹션 -->
-      <div
-        class="bg-gradient-to-r from-secondary-200 to-secondary-300 py-12 lg:py-16 px-[10.8%]"
-      >
-        <div class="flex justify-between items-start">
-          <!-- 왼쪽 컨텐츠 -->
-          <div class="flex-1">
-            <div class="flex items-center gap-3 mb-6">
-              <span
-                class="inline-flex items-center px-3 py-1 rounded-full bg-primary text-white text-caption font-medium"
-                >{{ productTypeName }}</span
-              >
-              <span
-                v-if="productDetail.joinMember"
-                class="inline-flex items-center px-3 py-1 rounded-full bg-fg-secondary-100 text-fg-primary text-caption"
-                >{{ productDetail.joinMember }}</span
-              >
-              <span
-                v-if="productDetail.joinWay"
-                class="inline-flex items-center px-3 py-1 rounded-full bg-fg-secondary-100 text-fg-primary text-caption"
-                >{{ productDetail.joinWay }}</span
-              >
-            </div>
-
-            <h1
-              class="text-title-lg lg:text-title-xl font-bold text-fg-primary mb-6"
-            >
-              {{ productDetail.name }}
-            </h1>
-
-            <div class="flex items-center gap-2 mb-8">
-              <span class="text-status-red">❤️</span>
-              <span class="text-footnote text-fg-secondary">
-                <strong class="text-fg-primary">{{
-                  productDetail.scrapCnt
-                }}</strong
-                >명이 관심갖고있음
-              </span>
-            </div>
-
-            <div
-              class="text-body02 text-fg-secondary leading-relaxed max-w-2xl"
-            >
-              제공되는 정보는 금융감독원
-              <strong class="text-fg-primary">{{ renewDate }}</strong
-              >일에 공시된 내용을 기반으로 작성되었으며, 금융상품 광고가
-              아닙니다. 실제 상품 가입 시점에 변동될 수 있으므로 상품 가입 시 꼭
-              다시 확인하시기 바랍니다.
-            </div>
-          </div>
-
-          <!-- 오른쪽 버튼들 -->
-          <div class="flex flex-col gap-3 ml-6 lg:ml-8 mt-auto">
-            <BaseButton
-              @click="addScrap(productDetail.id)"
-              label="🤍 관심상품"
-              size="sm"
-              variant="outline"
-              class="!w-auto !px-6 md:!px-8 lg:!px-10 !whitespace-nowrap !text-footnote md:!text-callout"
-            />
-            <BaseButton
-              @click="navigateToLink"
-              label="가입하기"
-              size="sm"
-              variant="filled"
-              class="!w-auto !px-6 md:!px-8 lg:!px-10 !whitespace-nowrap !text-footnote md:!text-callout"
-            />
-          </div>
-        </div>
-      </div>
+      <ProductHeader
+        :productDetail="productDetail"
+        :renewDate="renewDate"
+        @addScrap="handleAddScrap"
+        @navigate="navigateToLink"
+      />
 
       <!-- 메인 컨텐츠 섹션 -->
       <div class="space-y-4 py-8">
         <!-- 상품 주요 정보 카드 -->
-        <BaseCard
-          size="lg"
-          class="gap-8 flex justify-around items-start w-full p-8"
-        >
-          <!-- 최고금리 -->
-          <div class="text-center">
-            <h3 class="text-callout font-medium text-fg-secondary mb-3">
-              최고금리
-            </h3>
-            <p class="text-title-sm font-bold text-fg-primary">
-              {{ productDetail.primeRate }} %
-            </p>
-          </div>
-
-          <!-- 기본금리 -->
-          <div class="text-center">
-            <h3 class="text-callout font-medium text-fg-secondary mb-3">
-              기본금리
-            </h3>
-            <p class="text-title-sm font-bold text-fg-primary">
-              {{ productDetail.rate }} %
-            </p>
-          </div>
-
-          <!-- 기간 -->
-          <div class="text-center">
-            <h3 class="text-callout font-medium text-fg-secondary mb-3">
-              기간
-            </h3>
-            <p class="text-title-sm font-bold text-fg-primary">
-              {{ productDetail.saveTerm }} 개월
-            </p>
-          </div>
-
-          <!-- 나와의 적합도 -->
-          <div class="text-center">
-            <h3 class="text-callout font-medium text-fg-secondary mb-3">
-              나와의 적합도
-            </h3>
-            <!-- TODO: 이후 연결 -->
-            <p class="text-title-sm font-bold text-status-blue">50 %</p>
-          </div>
-        </BaseCard>
+        <ProductSummaryCard :productDetail="productDetail" />
 
         <!-- AI 한 줄 요약 -->
-        <div
-          class="bg-white rounded-xl shadow-card border border-1/2 border-secondary-200 p-8"
-        >
-          <h3 class="text-title-sm font-semibold text-fg-primary mb-6">
-            AI 한 줄 요약
-          </h3>
-          <div class="space-y-4">
-            <p class="text-body02 text-fg-secondary leading-relaxed">
-              <span class="text-callout font-medium text-status-blue"
-                >장점</span
-              >
-              : {{ productDetail.advantage }}
-            </p>
-            <p class="text-body02 text-fg-secondary leading-relaxed">
-              <span class="text-callout font-medium text-status-red">단점</span>
-              : {{ productDetail.disadvantage }}
-            </p>
-          </div>
-        </div>
+        <ProductAISummary :productDetail="productDetail" />
 
         <!-- 상품 정보 -->
-        <div
-          class="bg-white rounded-xl shadow-card border border-1/2 border-secondary-200 p-8"
-        >
-          <h3 class="text-title-sm font-semibold text-fg-primary mb-8">
-            상품 정보
-          </h3>
-
-          <div class="space-y-6">
-            <div class="flex items-start">
-              <span class="w-28 text-callout text-fg-secondary flex-shrink-0"
-                >최소 금액</span
-              >
-              <span class="text-body02 text-fg-primary font-medium"
-                >{{ productDetail.minimumCost }} 원 이상</span
-              >
-            </div>
-
-            <div class="flex items-start">
-              <span class="w-28 text-callout text-fg-secondary flex-shrink-0"
-                >운용 회사</span
-              >
-              <span class="text-body02 text-fg-primary font-medium">{{
-                productDetail.company
-              }}</span>
-            </div>
-
-            <div class="flex items-start" v-if="productDetail.joinMember">
-              <span class="w-28 text-callout text-fg-secondary flex-shrink-0"
-                >가입 대상</span
-              >
-              <span class="text-body02 text-fg-primary font-medium">{{
-                productDetail.joinMember
-              }}</span>
-            </div>
-
-            <div class="flex items-start" v-if="productDetail.joinWay">
-              <span class="w-28 text-callout text-fg-secondary flex-shrink-0"
-                >가입 방법</span
-              >
-              <span class="text-body02 text-fg-primary font-medium">{{
-                productDetail.joinWay
-              }}</span>
-            </div>
-
-            <div class="flex items-start" v-if="productDetail.caution">
-              <span class="w-28 text-callout text-fg-secondary flex-shrink-0"
-                >기타 사항</span
-              >
-              <span class="text-body02 text-fg-primary leading-relaxed">{{
-                productDetail.caution
-              }}</span>
-            </div>
-          </div>
-        </div>
+        <ProductInfo :productDetail="productDetail" />
 
         <!-- 수익 계산기 -->
         <ProfitCalculator
-          :rateOptions="calculatorProps.rateOptions || []"
-          :defaultRateIndex="0"
-          :minAmount="calculatorProps.minAmount || 10000"
+          :productDetail="productDetail"
           :maxAmount="100000000"
-          :period="calculatorProps.period || 12"
         />
       </div>
     </div>
